Select Group Input Module

Group of mutually dependent select menus for filtering data.frame's columns (like in Excel).

Usage

select_group_ui(
  id,
  params,
  label = NULL,
  btn_label = "Reset filters",
  inline = TRUE,
  vs_args = list()
)

select_group_server(id, data_r, vars_r)

Arguments

id

Module's id.

params

A list of parameters passed to each shinyWidgets::virtualSelectInput(), you can use :

inputId: mandatory, must correspond to variable name.

label: Display label for the control.

placeholder: Text to show when no options selected.

label

Character, global label on top of all labels.

btn_label

Character, reset button label.

inline

If TRUE (the default), select menus are horizontally positioned, otherwise vertically.

vs_args

Arguments passed to all shinyWidgets::virtualSelectInput() created.

data_r

Either a data.frame() or a shiny::reactive() function returning a data.frame (do not use parentheses).

vars_r

character, columns to use to create filters, must correspond to variables listed in params. Can be a shiny::reactive() function, but values must be included in the initial ones (in params).

Value

A shiny::reactive() function containing data filtered.

Examples

# Default -----------------------------------------------------------------

library(shiny)
library(datamods)
library(shinyWidgets)

ui <- fluidPage(
  # theme = bslib::bs_theme(version = 5L),
  fluidRow(
    column(
      width = 10, offset = 1,
      tags$h3("Filter data with select group module"),
      shinyWidgets::panel(
        select_group_ui(
          id = "my-filters",
          params = list(
            list(inputId = "Manufacturer", label = "Manufacturer:"),
            list(inputId = "Type", label = "Type:"),
            list(inputId = "AirBags", label = "AirBags:"),
            list(inputId = "DriveTrain", label = "DriveTrain:")
          )
        ),
        status = "primary"
      ),
      reactable::reactableOutput(outputId = "table")
    )
  )
)

server <- function(input, output, session) {
  res_mod <- select_group_server(
    id = "my-filters",
    data = reactive(MASS::Cars93),
    vars = reactive(c("Manufacturer", "Type", "AirBags", "DriveTrain"))
  )
  output$table <- reactable::renderReactable({
    reactable::reactable(res_mod())
  })
}

if (interactive())
  shinyApp(ui, server)