Import from all sources

Wrap all import modules into one, can be displayed inline or in a modal window..

Usage

import_ui(id, from = c("env", "file", "copypaste", "googlesheets", "url"))

import_server(
  id,
  validation_opts = NULL,
  allowed_status = c("OK", "Failed", "Error"),
  return_class = c("data.frame", "data.table", "tbl_df")
)

import_modal(id, from, title = "Import data", size = "l")

Arguments

id

Module's id

from

The import_ui & server to use, i.e. the method. There are 5 options to choose from. ("env", "file", "copypaste", "googlesheets", "url")

validation_opts

list of arguments passed to [validation_server().

allowed_status

Vector of statuses allowed to confirm dataset imported, if you want that all validation rules are successful before importing data use allowed_status = "OK".

return_class

Class of returned data: data.frame, data.table or tbl_df (tibble).

title

Modal window title.

size

Modal window size, default to "l" (large).

Value

UI: HTML tags that can be included in shiny's UI

Server: a list with three slots:

status: a reactive function returning the status: NULL, error or success.

name: a reactive function returning the name of the imported data as character.

data: a reactive function returning the imported data.frame.

Examples

library(shiny)
library(datamods)

ui <- fluidPage(
  fluidRow(
    column(
      width = 4,
      checkboxGroupInput(
        inputId = "from",
        label = "From",
        choices = c("env", "file", "copypaste", "googlesheets", "url"),
        selected = c("file", "copypaste")
      ),
      actionButton("launch_modal", "Launch modal window")
    ),
    column(
      width = 8,
      tags$b("Imported data:"),
      verbatimTextOutput(outputId = "name"),
      verbatimTextOutput(outputId = "data")
    )
  )
)

server <- function(input, output, session) {

  observeEvent(input$launch_modal, {
    req(input$from)
    import_modal(
      id = "myid",
      from = input$from,
      title = "Import data to be used in application"
    )
  })

  imported <- import_server("myid", return_class = "tbl_df")

  output$name <- renderPrint({
    req(imported$name())
    imported$name()
  })

  output$data <- renderPrint({
    req(imported$data())
    imported$data()
  })
}

if (interactive())
  shinyApp(ui, server)